Revert changes from hidden links
