+++ breathecode/events/actions.py
@@ -3,7 +3,7 @@ import re
 import logging
 
 from datetime import datetime, timedelta
-from breathecode.admissions.models import CohortTimeSlot
+from breathecode.admissions.models import Cohort, CohortTimeSlot
 from django.utils import timezone
 
 from .models import Organization, Venue, Event, Organizer
@@ -321,3 +321,27 @@ def get_cohort_description(timeslot: CohortTimeSlot) -> str:
     description += f'from {starting_hour} to {ending_hour}'
 
     return description.capitalize()
+
+
+def get_ical_cohort_description(item: Cohort):
+    description = ''
+    # description = f'{description}Url: {item.url}\n'
+
+    if item.name:
+        description = f'{description}Name: {item.name}\n'
+
+    if item.academy:
+        description = f'{description}Academy: {item.academy.name}\n'
+
+    if item.language:
+        description = f'{description}Language: {item.language.upper()}\n'
+
+    if item.private:
+        description = f'{description}Private: {"Yes" if item.private else "No"}\n'
+
+    if item.remote_available:
+        description = f'{description}Online: {"Yes" if item.remote_available else "No"}\n'
+
+    # TODO: add private url to meeting url
+
+    return description
